Factor out interpolation methods
